SMP: add 12 cores support. Test with nsim_hs5x_smp board (12 cores)

+++ include/zephyr/kernel/thread.h
@@ -120,7 +120,7 @@ struct _thread_base {
 
 #ifdef CONFIG_SCHED_CPU_MASK
 	/* "May run on" bits for each CPU */
-	uint8_t cpu_mask;
+	uint16_t cpu_mask;
 #endif
 
 	/* data returned by APIs */
